add tillClaim timer
