+++ src/pages/gd/Claim/ClaimBalance.tsx
@@ -4,7 +4,12 @@ import { ArrowButton, BalanceGD } from '@gooddollar/good-design'
 import { useSetChain } from '@web3-onboard/react'
 import { SupportedChains } from '@gooddollar/web3sdk-v2'
 import { ChainIdHex } from '../../../constants'
-import { Fraction } from '@uniswap/sdk-core'
+import { useClaim } from '@gooddollar/web3sdk-v2'
+import usePromise from 'hooks/usePromise'
+import { g$Price } from '@gooddollar/web3sdk'
+import { isToday, format } from 'date-fns'
+import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import { useClaiming } from 'hooks/useClaiming'
 
 const NextClaim = ({ time }: { time: string }) => {
     return (
@@ -15,18 +20,33 @@ const NextClaim = ({ time }: { time: string }) => {
 }
 
 const ClaimTimer = () => {
-    const timer = '21:00:00'
+    const { tillClaim } = useClaiming()
     return (
         <Box height="50" justifyContent="center" flexDirection="column" my="4">
             <Text fontFamily="subheading" fontSize="md" color="main">
                 Your next claim
             </Text>
-            <Text>{timer}</Text>
+            <Text>{tillClaim}</Text>
         </Box>
     )
 }
 
-export const ClaimBalance = ({ time, price }: { time: string | false; price: Fraction | undefined }) => {
+export const ClaimBalance = () => {
+    const { claimTime } = useClaim('everyBlock')
+    const { chainId } = useActiveWeb3React()
+    const [G$Price] = usePromise(async () => {
+        try {
+            const data = await g$Price()
+            return data.DAI
+        } catch {
+            return undefined
+        }
+    }, [chainId])
+
+    const formattedTime = useMemo(
+        () => (isToday(claimTime) ? 'today' : 'tomorrow') + ' ' + format(claimTime, 'hh aaa'),
+        [claimTime]
+    )
     const [{ connectedChain }, setChain] = useSetChain()
 
     const network = useMemo(
@@ -42,13 +62,13 @@ export const ClaimBalance = ({ time, price }: { time: string | false; price: Fra
     return (
         <View textAlign="center" display="flex" justifyContent="center" flexDirection="column" w="full" mb="4">
             <Box backgroundColor="goodWhite.100" borderRadius="15" p="1" w="full" h="34" justifyContent="center">
-                <NextClaim time={time || ''} />
+                <NextClaim time={formattedTime || ''} />
             </Box>
 
             <ClaimTimer />
             <Box borderWidth="1" borderColor="borderGrey" width="90%" alignSelf="center" my="2" />
             <Box>
-                <BalanceGD gdPrice={price} />
+                <BalanceGD gdPrice={G$Price} />
             </Box>
             <Box alignItems="center">
                 <ArrowButton
